requirements.txt
aiohttp==3.8.4
pycryptodome==3.17
